delete acquires upload lock and deletes only if metadata blob checksum exists

+++ src/sentry/tasks/files.py
@@ -10,6 +10,12 @@ from sentry.tasks.deletion import MAX_RETRIES
     default_retry_delay=60 * 5,
     max_retries=MAX_RETRIES
 )
-def delete_file(path, **kwargs):
-    from sentry.models.file import get_storage
-    get_storage().delete(path)
+def delete_file(path, checksum, **kwargs):
+    from sentry.models.file import get_storage, FileBlob
+    from sentry.app import locks
+    from sentry.utils.retries import TimedRetryPolicy
+
+    lock = locks.get('fileblob:upload:{}'.format(checksum), duration=60 * 10)
+    with TimedRetryPolicy(60)(lock.acquire):
+        if FileBlob.objects.filter(checksum=checksum).exists():
+            get_storage().delete(path)
